Internal API

Note

This is the documentation of PartitionedParallelSampling's internal API, it is subject to change without deprecation.

Types

PartitionedParallelSampling.KDTreePartitioning

PartitionedParallelSampling.SpacePartTree

Functions and macros

PartitionedParallelSampling.partition_space

Documentation

PartitionedParallelSampling.KDTreePartitioning — Type

struct KDTreePartitioning <: SpacePartitioningAlgorithm

Internal, not part of stable public API.

K-D binary space partitioning algorithm. By default, all parameters are considered for partitioning. Partition parameters can be specified manually by using partition_dims argument. By default, bounds of the partitioning tree are extended to those given by prior. This can be changed by setting extend_bounds = false.

Constructors:

KDTreePartitioning(; fields...)

Fields:

partition_dims::Union{Symbol, Vector{Int64}}

Default: :auto

extend_bounds::Bool

Default: true

PartitionedParallelSampling.SpacePartTree — Type

SpacePartTree

Internal, not part of stable public API.

The structure stores a partitioning tree generated by any SpacePartitioningAlgorithm. Variables:

terminated_leaf : true if the tree node is terminal, false otherwise.

bounds : Low and high bound of the tree leaf.

left_child : The left child of the tree, missing is the node is terminal.

right_child : The right child of the tree, missing is the node is terminal.

cut_axis : Axis along which cut is performed, missing if the node is terminal.

cut_coordinate : Coordinate at which a cut is performed, missing if the node is terminal.

cost : The sum of the cost functions over leaves.

cost_part : The cost function of the current leaf, missing if the node is terminal..

PartitionedParallelSampling.partition_space — Function

partition_space(
	samples::DensitySampleVector,
	npartitions::Integer,
	algorithm::KDTreePartitioning
)

Internal, not part of stable public API.

The function generates a space partition tree with the number of partitions given by npartitions, using KDTreePartitioning algorithm. Exploration samples are given by samples. The output contains SpacePartTree and the values of the cost function.